move `HlsSegment` to `model` package
